# openapi.yaml
openapi: 3.1.1
info:
  title: Engineering Thesis API
  version: 1.0.0
servers:
  - url: http://localhost:8080
components:
  securitySchemes:
    JwtBearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    GetEnterpriseResponse:
      type: object
      required: [name, description]
      properties:
        name:
          type: string
          example: The Best Barbershop
        description:
          type: string
          example: The new best barbershop in town.
        location:
          $ref: "#/components/schemas/Location"
...
    IanaTimeZone:
      type: string
      pattern: "^[A-Za-z_]+/[A-Za-z_]+$"
      example: "Europe/Warsaw"
...
    Location:
      type: object
      required: [address, longitude, latitude]
      properties:
        address:
          type: string
          example: ul. Nowowiejska 15/19, Warsaw, Poland
        longitude:
          type: number
          format: double
          example: 12.3248
        latitude:
          type: number
          format: double
          example: -67.224
...
    LocalDatetime:
      type: string
      pattern: "^\\d{4}-\\d{2}-\\d{2}T\\d{2}:\\d{2}(:\\d{2})?$"
      example: "2025-11-05T10:30:00"
    GetCustomerLandingPageScheduledAppointmentResponseItem:
      type: object
      required:
        [
          serviceName,
          enterpriseName,
          address,
          startDatetimeServiceLocal,
          endDatetimeServiceLocal,
          timezone,
          price,
        ]
      properties:
        serviceName:
          type: string
        enterpriseName:
          type: string
        address:
          type: string
        startDatetimeServiceLocal:
          $ref: "#/components/schemas/LocalDatetime"
        endDatetimeServiceLocal:
          $ref: "#/components/schemas/LocalDatetime"
        timezone:
          $ref: "#/components/schemas/IanaTimeZone"
        price:
          type: number
    GetCustomerLandingPagePendingAppointmentResponseItem:
      type: object
      required:
        [
          serviceName,
          enterpriseName,
          address,
          startDatetimeServiceLocal,
          endDatetimeServiceLocal,
          timezone,
          price,
        ]
      properties:
        serviceName:
          type: string
        enterpriseName:
          type: string
        address:
          type: string
        startDatetimeServiceLocal:
          $ref: "#/components/schemas/LocalDatetime"
        endDatetimeServiceLocal:
          $ref: "#/components/schemas/LocalDatetime"
        timezone:
          $ref: "#/components/schemas/IanaTimeZone"
        price:
          type: number
    GetCustomerLandingPageRejectedAppointmentResponseItem:
      type: object
      required:
        [
          serviceName,
          enterpriseName,
          address,
          startDatetimeServiceLocal,
          endDatetimeServiceLocal,
          timezone,
          price,
          rejectionMessage,
        ]
      properties:
        serviceName:
          type: string
        enterpriseName:
          type: string
        address:
          type: string
        startDatetimeServiceLocal:
          $ref: "#/components/schemas/LocalDatetime"
        endDatetimeServiceLocal:
          $ref: "#/components/schemas/LocalDatetime"
        timezone:
          $ref: "#/components/schemas/IanaTimeZone"
        price:
          type: number
        rejectionMessage:
          type: string
...
  responses:
...
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            type: object
            required:
              - message
            properties:
              message:
                type: string
                example: Unauthorized
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            type: object
            required:
              - message
            properties:
              message:
                type: string
                example: Bad request
    Forbidden:
      description: Forbidden
      content:
        application/json:
          schema:
            type: object
            required:
              - message
            properties:
              message:
                type: string
                example: Forbidden request
...
    InternalServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            type: object
            required:
              - message
            properties:
              message:
                type: string
                example: Internal server error
...
paths:
...
  /enterprises/{enterpriseId}:
    get:
      tags: [Enterprises]
      operationId: getEnterprise
      security:
        - JwtBearerAuth: []
      parameters:
        - name: enterpriseId
          in: path
          required: true
          description: Enterprise id
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: *OK_DESCRIPTION
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GetEnterpriseResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "500":
          $ref: "#/components/responses/InternalServerError"
...
  /appointments/my-future-scheduled-appointments:
    get:
      tags: [Appointments]
      operationId: getMyFutureScheduledAppointments
      security:
        - JwtBearerAuth: []
      responses:
        "200":
          description: *OK_DESCRIPTION
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/GetCustomerLandingPageScheduledAppointmentResponseItem"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /appointments/my-pending-appointments:
    get:
      tags: [Appointments]
      operationId: getMyPendingAppointments
      security:
        - JwtBearerAuth: []
      responses:
        "200":
          description: *OK_DESCRIPTION
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/GetCustomerLandingPagePendingAppointmentResponseItem"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /appointments/my-rejected-appointments:
    get:
      tags: [Appointments]
      operationId: getMyRejectedAppointments
      security:
        - JwtBearerAuth: []
      responses:
        "200":
          description: *OK_DESCRIPTION
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/GetCustomerLandingPageRejectedAppointmentResponseItem"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /appointments/my-past-appointments:
    get:
      tags: [Appointments]
      operationId: getMyPastScheduledAppointments
      security:
        - JwtBearerAuth: []
      responses:
        "200":
          description: *OK_DESCRIPTION
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/GetCustomerLandingPageScheduledAppointmentResponseItem"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "500":
          $ref: "#/components/responses/InternalServerError"
...